Struct elastic::client::prelude::RequestParams

pub struct RequestParams { /* fields omitted */ }

Parameters for a single REST API request.

The RequestParams struct allows you to set headers and url parameters for a given request. By default, the ContentType::json header will always be added. Url parameters are added as simple key-value pairs, and serialised by rust-url.

Examples

With default query parameters:

let params = RequestParams::default();

With a custom base url:

let params = RequestParams::new("http://mybaseurl:9200");

With custom headers:

let params = RequestParams::default()
    .header(Authorization("let me in".to_owned()));

With url query parameters:

let params = RequestParams::default()
    .url_param("pretty", true)
    .url_param("q", "*");

Methods

impl RequestParams

pub fn from_parts<T: Into<NodeAddress>>(
base_url: T,
inner: PreRequestParams
) -> Self

Create a container for request parameters from a base url and pre request parameters.

pub fn new<T: Into<NodeAddress>>(base_url: T) -> Self

Create a new container for request parameters.

This method takes a fully-qualified url for the Elasticsearch node. It will also set the Content-Type header to application/json.

pub fn base_url<T: Into<NodeAddress>>(self, base_url: T) -> Self

Set the base url for the Elasticsearch node.

pub fn url_param<T: ToString>(self, key: &'static str, value: T) -> Self

Set a url param value.

These parameters are added as query parameters to request urls.

pub fn header<H>(self, header: H) -> Self where
H: Header + Clone,

Set a request header.

pub fn get_base_url(&self) -> &str

Get the base url.

pub fn get_headers(&self) -> Headers

Create a new Headers structure, and thread it through the configuration functions.

pub fn get_url_qry(&self) -> (usize, Option<String>)

Get the url query params as a formatted string.

Follows the application/x-www-form-urlencoded format. This method returns the length of the query string and an optional value. If the value is None, then the length will be 0.

pub fn split(self) -> (NodeAddress, PreRequestParams)

Split the request parameters into its parts.