Module elastic::client::requests

Request types for the Elasticsearch REST API.

Reexports

pub use self::params::*;

pub use self::endpoints::*;

Modules

endpoints
REST API endpoints.

params
Common url params like Id and Index.

Structs

HttpRequest
A general request type that all endpoints can be converted into.

Url
A wrapper around an owned or borrowed url.

Enums

HttpMethod
A standard HTTP verb.

Traits

IntoBody
A type that can be converted into a reqwest::Body.

TryForDoc
A trait for converting a document into a request.

TryForMapping
A trait for converting a document mapping mapping into a request.

Functions

empty_body
A convenience method for a default, empty body. This method doesn't allocate.