Type Definition elastic::client::requests::bulk::BulkRequestBuilder

type BulkRequestBuilder<TSender, TBody, TResponse> = RequestBuilder<TSender, BulkRequestInner<TBody, TResponse>>;

A bulk request builder that can be configured before sending.

Call Client.bulk to get a BulkRequestBuilder. The send method will either send the request synchronously or asynchronously, depending on the Client it was created from.

Call Client.bulk_stream to get a BulkRequestBuilder that can be used to stream bulk operations asynchronously.

Methods

impl<TSender, TBody, TResponse> BulkRequestBuilder<TSender, TBody, TResponse> where
TSender: Sender,

Builder methods

Configure a BulkRequestBuilder before sending it.

pub fn ty<I>(self, ty: I) -> Self where
I: Into<Type<'static>>,

Set the default type for the bulk request.

If an operation doesn't specify a type, then it will default to the supplied value here.

Deferred errors

Calling ty without also calling index will result in an error when sending the request.

pub fn index<I>(self, index: I) -> Self where
I: Into<Index<'static>>,

Set the default index for the bulk request.

If an operation doesn't specify an index, then it will default to the supplied value here.

pub fn response_index<I>(
self
) -> BulkRequestBuilder<TSender, TBody, TResponse::WithNewIndex> where
TResponse: ChangeIndex<I>,

Set the type used to deserialize the index field on the response.

Sometimes a bulk response will use the same index value many times. To avoid allocating a lot of individual strings, the type used to deserialize the field can be changed. string_cache::DefaultAtom or a custom enum can be effective ways to reduce allocations in large bulk responses.

pub fn response_ty<I>(
self
) -> BulkRequestBuilder<TSender, TBody, TResponse::WithNewType> where
TResponse: ChangeType<I>,

Set the type used to deserialize the type field on the response.

Sometimes a bulk response will use the same type value many times. To avoid allocating a lot of individual strings, the type used to deserialize the field can be changed. string_cache::DefaultAtom or a custom enum can be effective ways to reduce allocations in large bulk responses.

pub fn response_id<I>(
self
) -> BulkRequestBuilder<TSender, TBody, TResponse::WithNewId> where
TResponse: ChangeId<I>,

Set the type used to deserialize the id field on the response.

It's less likely that id fields in bulk responses will be repeated, but they're probably short. To avoid allocating a lot of individual strings, the type used to deserialize the field can be changed. inlinable_string::InlinableString can be an effective way to recude allocation in large bulk responses.

impl<TSender, TBody, TIndex, TType, TId> BulkRequestBuilder<TSender, TBody, BulkResponse<TIndex, TType, TId>> where
TSender: Sender,

pub fn errors_only(
self
) -> BulkRequestBuilder<TSender, TBody, BulkErrorsResponse<TIndex, TType, TId>>

Only deserialize failed bulk operations in the response.

Elasticsearch returns a response that's proportional in size to the number of operations in the request. If you only care about failures then it can be more efficient to ignore the common case where operations succeed.

impl<TSender, TBody, TResponse> BulkRequestBuilder<TSender, TBody, TResponse> where
TSender: Sender,
TBody: BulkBody,

pub fn push<TDocument, TOperation>(self, op: TOperation) -> Self where
TOperation: Into<BulkOperation<TDocument>>,
TDocument: Serialize,

Push an operation onto the bulk request.

Deferred errors

If the document can't be serialized then sending the request will return an error.

pub fn extend<TIter, TDocument>(self, iter: TIter) -> Self where
TIter: IntoIterator<Item = BulkOperation<TDocument>>,
TDocument: Serialize,

Push a collection of operations onto the bulk request.

Deferred errors

If any documents can't be serialized then sending the request will return an error.

impl<TDocument, TResponse> BulkRequestBuilder<AsyncSender, Streamed<TDocument>, TResponse>

Stream builder methods

Configure a SearchRequestBuilder before sending it.

pub fn timeout(self, timeout: Duration) -> Self

Specify a timeout for filling up the request buffer.

This parameter can be used to control the miminum frequency of bulk requests. If the timeout expires before the buffer is full then a bulk request will be sent with whatever data was written. The timeout isn't restarted when operations are pushed.

pub fn body_size_bytes(self, body_size: usize) -> Self

Specify a maximum request size in bytes.

This parameter can be used to control the maximum size of a single bulk request emitted. Operations will be appended to the request until the body_size is reached.

pub fn build(
self
) -> (BulkSender<TDocument, TResponse>, BulkReceiver<TResponse>)

Create a channel for streaming bulk operations.

This will return a channel with a [BulkSender] and [BulkReceiver] pair. Push operations into the sender. Once an internal buffer is full, or a timeout expires then the bulk request will be sent. Responses can be pulled by the receiver.

Examples

TODO

impl<TBody, TResponse> BulkRequestBuilder<SyncSender, TBody, TResponse> where
TBody: Into<SyncBody> + BulkBody + 'static,
TResponse: DeserializeOwned + IsOk + 'static,

Send synchronously

pub fn send(self) -> Result<TResponse, Error>

Send a BulkRequestBuilder synchronously using a SyncClient.

This will block the current thread until a response arrives and is deserialised.

Examples

Send a bulk request to index some documents:

let ops = (0..1000)
    .into_iter()
    .map(|i| bulk_index(MyType {
            id: i,
            title: "some string value".into()
        })
        .id(i));

let response = client.bulk()
                     .index("myindex")
                     .ty(MyType::name())
                     .extend(ops)
                     .send()?;

for op in response {
    match op {
        Ok(op) => println!("ok: {:?}", op),
        Err(op) => println!("err: {:?}", op),
    }
}

impl<TBody, TResponse> BulkRequestBuilder<AsyncSender, TBody, TResponse> where
TBody: Into<AsyncBody> + BulkBody + Send + 'static,
TResponse: DeserializeOwned + IsOk + Send + 'static,

Send asynchronously

pub fn send(self) -> Pending<TResponse>

Send a BulkRequestBuilder asynchronously using an AsyncClient.

This will return a future that will resolve to the deserialised search response.

Examples

Send a bulk request to index some documents:

let ops = (0..1000)
    .into_iter()
    .map(|i| bulk_index(MyType {
            id: i,
            title: "some string value".into()
        })
        .id(i));

let future = client.bulk()
                     .index("myindex")
                     .ty(MyType::name())
                     .extend(ops)
                     .send();

future.and_then(|response| {
    for op in response {
        match op {
            Ok(op) => println!("ok: {:?}", op),
            Err(op) => println!("err: {:?}", op),
        }
    }

    Ok(())
});