Struct elastic::client::requests::document_update::ScriptBuilder

pub struct ScriptBuilder<TParams> { /* fields omitted */ }

A builder for an update script that can be configured before sending.

Methods

impl ScriptBuilder<Map<String, Value>>

pub fn new<TScript>(source: TScript) -> Self where
TScript: ToString,

Create a new script builder using the given source.

pub fn param<TKey, TValue>(self, key: TKey, value: TValue) -> Self where
TKey: ToString,
TValue: Into<Value>,

Set a script parameter.

impl<TParams> ScriptBuilder<TParams>

pub fn lang<TLang>(self, lang: Option<TLang>) -> Self where
TLang: ToString,

Set the language for the update script.

pub fn params<TNewParams>(self, params: TNewParams) -> ScriptBuilder<TNewParams>

Specify a new set of parameters for the update script.