Module elastic::client::requests::params

Common url params like Id and Index.

The parameter types are basically just a wrapper around a maybe owned string. They can all be constructed from a String or an &str, but some parameters may have other implementations in the future.

Structs

Alias

Feature

Fields

Id

Index

IndexMetric

Lang

Metric

Name

NewIndex

NodeId

Repository

ScrollId

Snapshot

Target

TaskId

ThreadPoolPatterns

Type