Module elastic::client::responses

Response types for the Elasticsearch REST API.

Modules

parse
Utility types for response parsing.

Structs

AggregationIterator
Aggregator that traverses the results from Elasticsearch's Aggregations and returns a result row by row in a table-styled fashion.

Aggregations
Type Struct to hold a generic serde_json::Value tree of the Aggregation results.

Hit

Hits
Struct to hold the search's Hits, serializable to type T or serde_json::Value

HttpResponse
A raw HTTP response with enough information to parse a concrete type from it.

Shards

Type Definitions

GetResponse
A generic Get Document API response.

SearchResponse
A generic Search API response.