Struct elastic::client::responses::HttpResponse

pub struct HttpResponse<R> { /* fields omitted */ }

A raw HTTP response with enough information to parse a concrete type from it.

Methods

impl<R> HttpResponse<R> where
R: Read,

fn response<T, F>(self, is_ok: F) -> Result<T, ResponseError> where
F: Fn(UnbufferedResponse<R>) -> Result<MaybeOkResponse<R>, Error>,
T: Deserialize,

Get the response body from JSON.

This method takes a closure that determines whether the result is successful. If the MaybeOkResponse is ok, then the body will be returned as Ok(T). Otherwise the body will be returned as Err(ApiError).

impl<R> HttpResponse<R>

fn new(status: u16, body: R) -> HttpResponse<R>

Create a new HTTP response from the given status code and body.

fn status(&self) -> u16

Get the status code.