Type Definition elastic::client::responses::SearchResponse

type SearchResponse<T> = SearchResponseOf<Hit<T>>;

A generic Search API response.

For responses that contain multiple document types, use SearchResponse<serde_json::Value>.

This type won't parse if you've applied any response filters. If you need to tweak the shape of the search response you can define your own response type and implement FromResponse for it. See the parse mod for more details.