Struct elastic::client::sender::AsyncClientBuilder

pub struct AsyncClientBuilder { /* fields omitted */ }

A builder for an asynchronous client.

Methods

impl AsyncClientBuilder

pub fn new() -> Self

Create a new client builder.

By default, a client constructed by this builder will:

Send requests to localhost:9200

Not deserialise repsonses on a cpu pool

Not use any authentication

Not use TLS

pub fn from_params(params: PreRequestParams) -> Self

Create a new client builder with the given default request parameters.

pub fn static_node<S>(self, node: S) -> Self where
S: Into<NodeAddress>,

Specify a static node nodes to send requests to.

pub fn static_nodes<I, S>(self, nodes: I) -> Self where
I: IntoIterator<Item = S>,
S: Into<NodeAddress>,

Specify a set of static node nodes to load balance requests on.

pub fn sniff_nodes<I>(self, builder: I) -> Self where
I: Into<SniffedNodesBuilder>,

Specify a node address to sniff other nodes in the cluster from.

Examples

Use a given base url for sniffing the cluster's node addresses from:

let builder = AsyncClientBuilder::new()
    .sniff_nodes("http://localhost:9200");

pub fn sniff_nodes_fluent<I, F>(self, address: I, builder: F) -> Self where
I: Into<NodeAddress>,
F: Fn(SniffedNodesBuilder) -> SniffedNodesBuilder + 'static,

Specify a node address to sniff other nodes in the cluster from.

Examples

Use a given base url for sniffing the cluster's node addresses from and specify a minimum duration to wait before refreshing:

let builder = AsyncClientBuilder::new()
    .sniff_nodes_fluent("http://localhost:9200", |n| n
        .wait(Duration::from_secs(90)));

pub fn params_fluent<F>(self, builder: F) -> Self where
F: Fn(PreRequestParams) -> PreRequestParams + 'static,

Specify default request parameters.

Examples

Require all responses use pretty-printing:

let builder = AsyncClientBuilder::new()
    .params_fluent(|p| p
        .url_param("pretty", true));

Add an authorization header:

use elastic::http::header::Authorization;

let builder = AsyncClientBuilder::new()
    .params_fluent(|p| p
        .header(Authorization("let me in".to_owned())));

pub fn params<I>(self, params: I) -> Self where
I: Into<PreRequestParams>,

Specify default request parameters.

Examples

Require all responses use pretty-printing:

let builder = AsyncClientBuilder::new()
    .params(PreRequestParams::new()
        .url_param("pretty", true));

Add an authorization header:

use elastic::http::header::Authorization;

let builder = AsyncClientBuilder::new()
    .params(PreRequestParams::new()
        .header(Authorization("let me in".to_owned())));

pub fn serde_pool<P>(self, serde_pool: P) -> Self where
P: Into<Option<CpuPool>>,

Use the given CpuPool for serialising and deserialising responses.

If the pool is None then responses will be serialised and deserialised on the same thread as the io Core.

Examples

Use a cpu pool to serialise and deserialise responses:

let pool = CpuPool::new(4);

let builder = AsyncClientBuilder::new().serde_pool(pool);

pub fn pre_send_raw<F>(self, pre_send: F) -> Self where
F: Fn(&mut AsyncHttpRequest) -> Box<Future<Item = (), Error = Box<StdError + Send + Sync>>> + 'static,

Specify a function to tweak a raw request before sending.

This function will be applied to all outgoing requests and gives you the chance to perform operations the require the complete raw request, such as request singing. Prefer the params method on the client or individual requests where possible.

pub fn build<TIntoHttp>(self, client: TIntoHttp) -> Result<AsyncClient, Error> where
TIntoHttp: IntoAsyncHttpClient,

Construct an AsyncClient from this builder.

The build method accepts any type that can be used to construct a http client from.

Examples

Build with an asynchronous Handle. This will build an AsyncClient with a default underlying AsyncHttpClient using the handle.

let mut core = Core::new()?;

let builder = AsyncClientBuilder::new().build(&core.handle());

Build with a given AsyncHttpClient.

let mut core = Core::new()?;
let client = Client::new(&core.handle());

let builder = AsyncClientBuilder::new().build(client);