Trait elastic::client::sender::Sender

pub trait Sender: Sealed + Clone {
    type Body;
    type Response;
    type Params;
    fn send<TRequest, TParams, TBody>(
        &self, 
        request: SendableRequest<TRequest, TParams, TBody>
    ) -> Self::Response
    where
        TRequest: Into<HttpRequest<'static, TBody>>,
        TBody: Into<Self::Body> + 'static,
        TParams: Into<Self::Params> + 'static;
}

Represents a type that can send a request.

You probably don't need to touch this trait directly. See the Client type for making requests.

Even though the Sender trait is quite generic, it's not really designed to be implemented externally. The request builders expect there to be 2 concrete implementations of Sender, namely SyncSender and AsyncSender. The real purpose is to make it possible to share builders so the sync and async APIs don't diverge.

At some point in the future though this may be made more generic so you could reasonably plug your own Senders in to elastic.

Associated Types

type Body

type Response

type Params

Required Methods

fn send<TRequest, TParams, TBody>(
&self,
request: SendableRequest<TRequest, TParams, TBody>
) -> Self::Response where
TRequest: Into<HttpRequest<'static, TBody>>,
TBody: Into<Self::Body> + 'static,
TParams: Into<Self::Params> + 'static,