Struct elastic::client::DocumentClient

pub struct DocumentClient<TSender, TDocument> { /* fields omitted */ }

A Client for a specific document type.

Methods

impl<TSender, TDocument> DocumentClient<TSender, TDocument> where
TSender: Sender,

Search request

pub fn search(self) -> SearchRequestBuilder<TSender, TDocument, DefaultBody> where
TDocument: DeserializeOwned + DocumentType,

Create a SearchRequestBuilder with this Client that can be configured before sending.

The index and type parameters will be inferred from the document type.

For more details, see:

builder methods

send synchronously

send asynchronously

Examples

Run a simple Query String query for a DocumentType called MyType:

let query = "a query string";

let response = client.document::<MyType>()
                     .search()
                     .body(json!({
                         "query": {
                             "query_string": {
                                 "query": query
                             }
                         }
                     }))
                     .send()?;

// Iterate through the hits (of type `MyType`)
for hit in response.hits() {
    println!("{:?}", hit);
}

For more details on document types and mapping, see the types module.

impl<TSender, TDocument> DocumentClient<TSender, TDocument> where
TSender: Sender,

Get document request

pub fn get<TId>(self, id: TId) -> GetRequestBuilder<TSender, TDocument> where
TId: Into<Id<'static>>,
TDocument: DeserializeOwned + DocumentType + StaticIndex + StaticType,

Create a GetRequestBuilder with this Client that can be configured before sending.

For more details, see:

builder methods

send synchronously

send asynchronously

Examples

Get a DocumentType called MyType with an id of 1:

let response = client.document::<MyType>()
                     .get(1)
                     .send()?;

For more details on document types, see the types module.

pub fn get_raw<TIndex, TId>(
self,
index: TIndex,
id: TId
) -> GetRequestBuilder<TSender, TDocument> where
TIndex: Into<Index<'static>>,
TId: Into<Id<'static>>,
TDocument: DeserializeOwned,

Create a GetRequestBuilder with this Client that can be configured before sending.

For more details, see:

builder methods

send synchronously

send asynchronously

Examples

Get a document as a serde_json::Value:

let response = client.document::<Value>()
                     .get_raw("myindex", 1)
                     .send()?;

impl<TSender, TDocument> DocumentClient<TSender, TDocument> where
TSender: Sender,

Index document request

pub fn index(self, doc: TDocument) -> IndexRequestBuilder<TSender, TDocument> where
TDocument: Serialize + DocumentType,

Create a IndexRequestBuilder with this Client that can be configured before sending.

For more details, see:

builder methods

send synchronously

send asynchronously

Examples

Index a DocumentType called MyType with an id of 1:

#[derive(Serialize, Deserialize, ElasticType)]
struct MyType {
    #[elastic(id)]
     pub id: String,
     pub title: String,
     pub timestamp: Date<DefaultDateMapping>
}

let doc = MyType {
    id: "1".to_owned(),
    title: "A title".to_owned(),
    timestamp: Date::now()
};

let response = client.document()
                     .index(doc)
                     .send()?;
    
assert!(response.created());

For more details on document types and mapping, see the types module.

pub fn index_raw<TIndex>(
self,
index: TIndex,
doc: TDocument
) -> IndexRequestBuilder<TSender, TDocument> where
TIndex: Into<Index<'static>>,
TDocument: Serialize,

Create a IndexRequestBuilder with this Client that can be configured before sending.

For more details, see:

builder methods

send synchronously

send asynchronously

Examples

Index a DocumentType called MyType with an id of 1:

let doc_id = 123;
let doc = json!({
    "id": doc_id,
    "title": "A document"
});

let response = client.document()
                     .index_raw("myindex", doc)
                     .id(doc_id)
                     .send()?;
    
assert!(response.created());

For more details on document types and mapping, see the types module.

impl<TSender, TDocument> DocumentClient<TSender, TDocument> where
TSender: Sender,

Update document request

pub fn update<TId>(
self,
id: TId
) -> UpdateRequestBuilder<TSender, Doc<TDocument>> where
TId: Into<Id<'static>>,
TDocument: DocumentType + StaticIndex + StaticType,

Create an UpdateRequestBuilder with this Client that can be configured before sending.

For more details, see:

builder methods

send synchronously

send asynchronously

Examples

Update a DocumentType called MyType with an id of 1 using a new document value:

let response = client.document::<MyType>()
                     .update(1)
                     .doc(new_doc)
                     .send()?;

assert!(response.updated());

The document doesn't necessarily need to be of the same type so partial updates can be made. Be careful though because this can lead to unexpected results or runtime errors if the document types don't align:

let response = client.document::<MyType>()
                     .update(1)
                     .doc(json!({
                         "title": "New Title"
                     }))
                     .send()?;

assert!(response.updated());

Documents can also be updated using a script:

let response = client.document::<MyType>()
                     .update(1)
                     .script(r#"ctx._source.title = "New Title""#)
                     .send()?;

assert!(response.updated());

Scripts can be configured with parameters:

let response = client.document::<MyType>()
                     .update(1)
                     .script_fluent("ctx._source.title = params.newTitle", |script| script
                        .param("newTitle", "New Title"))
                     .send()?;

assert!(response.updated());

impl<TSender> DocumentClient<TSender, ()> where
TSender: Sender,

pub fn update_raw<TIndex, TId>(
self,
index: TIndex,
id: TId
) -> UpdateRequestBuilder<TSender, Doc<()>> where
TIndex: Into<Index<'static>>,
TId: Into<Id<'static>>,

Create an [UpdateRequestBuilder][UpdateRequestBuilder] with this Client that can be configured before sending.

For more details, see:

[builder methods][builder-methods]

[send synchronously][send-sync]

[send asynchronously][send-async]

Examples

Update a document with an id of 1 using a new document value:

let response = client.document()
                     .update_raw("myindex", 1)
                     .doc(json!({
                         "title": "New Title"
                     }))
                     .send()?;

assert!(response.updated());

impl<TSender, TDocument> DocumentClient<TSender, TDocument> where
TSender: Sender,

Delete document request

pub fn delete<TId>(self, id: TId) -> DeleteRequestBuilder<TSender, TDocument> where
TId: Into<Id<'static>>,
TDocument: DocumentType + StaticIndex + StaticType,

Create a DeleteRequestBuilder with this Client that can be configured before sending.

For more details, see:

builder methods

send synchronously

send asynchronously

Examples

Delete a DocumentType called MyType with an id of 1:

let response = client.document::<MyType>()
                     .delete(1)
                     .send()?;

assert!(response.deleted());

impl<TSender> DocumentClient<TSender, ()> where
TSender: Sender,

pub fn delete_raw<TIndex, TId>(
self,
index: TIndex,
id: TId
) -> DeleteRequestBuilder<TSender, ()> where
TIndex: Into<Index<'static>>,
TId: Into<Id<'static>>,

Create a DeleteRequestBuilder with this Client that can be configured before sending.

For more details, see:

builder methods

send synchronously

send asynchronously

Examples

Delete a DocumentType called MyType with an id of 1:

let response = client.document()
                     .delete_raw("myindex", 1)
                     .send()?;

assert!(response.deleted());

impl<TSender, TDocument> DocumentClient<TSender, TDocument> where
TSender: Sender,

Put mapping request

pub fn put_mapping(self) -> PutMappingRequestBuilder<TSender, TDocument> where
TDocument: DocumentType + StaticIndex + StaticType,

Create a PutMappingRequestBuilder with this Client that can be configured before sending.

For more details, see:

builder methods

send synchronously

send asynchronously

Examples

Put the document mapping for a DocumentType called MyType:

let response = client.document::<MyType>()
                     .put_mapping()
                     .send()?;

assert!(response.acknowledged());

For more details on document types and mapping, see the types module.