Struct elastic::client::IndexClient

pub struct IndexClient<TSender> { /* fields omitted */ }

A Client for a specific index.

Methods

impl<TSender> IndexClient<TSender> where
TSender: Sender,

Create index request

pub fn create(self) -> IndexCreateRequestBuilder<TSender, DefaultBody>

Create an IndexCreateRequestBuilder with this Client that can be configured before sending.

For more details, see:

builder methods

send synchronously

send asynchronously

Examples

Create an index called myindex:

let my_index = index("myindex");

let response = client.index(my_index).create().send()?;

assert!(response.acknowledged());

Create an index with settings and document mappings for a DocumentType called MyType:

let body = json!({
    "settings": {
        "index": {
            "number_of_shards": 3,
            "number_of_replicas": 2
        }
    },
    "mappings": {
        MyType::static_index(): MyType::index_mapping()
    }
});

let response = client.index("myindex")
                     .create()
                     .body(body.to_string())
                     .send()?;

assert!(response.acknowledged());

For more details on document types and mapping, see the types module.

impl<TSender> IndexClient<TSender> where
TSender: Sender,

Open index request

pub fn open(self) -> IndexOpenRequestBuilder<TSender>

Open an IndexOpenRequestBuilder with this Client that can be configured before sending.

For more details, see:

send synchronously

send asynchronously

Examples

Open an index called myindex:

let response = client.index("myindex").open().send()?;

assert!(response.acknowledged());

impl<TSender> IndexClient<TSender> where
TSender: Sender,

Close index request

pub fn close(self) -> IndexCloseRequestBuilder<TSender>

Create an IndexCloseRequestBuilder with this Client that can be configured before sending.

For more details, see:

send synchronously

send asynchronously

Examples

Close an index called myindex:

let response = client.index("myindex").close().send()?;

assert!(response.acknowledged());

impl<TSender> IndexClient<TSender> where
TSender: Sender,

Delete index request

pub fn delete(self) -> IndexDeleteRequestBuilder<TSender>

Delete a IndexDeleteRequestBuilder with this Client that can be configured before sending.

For more details, see:

send synchronously

send asynchronously

Examples

Delete an index called myindex:

let response = client.index("myindex").delete().send()?;

assert!(response.acknowledged());

impl<TSender> IndexClient<TSender> where
TSender: Sender,

Index exists request

pub fn exists(self) -> IndexExistsRequestBuilder<TSender>

Open an IndexExistsRequestBuilder with this Client that can be configured before sending.

For more details, see:

send synchronously

send asynchronously

Examples

Check whether an index called myindex exists:

let response = client.index("myindex").exists().send()?;

assert!(response.exists());