Struct elastic::client::PreRequestParams

pub struct PreRequestParams { /* fields omitted */ }

An incomplete set of request parameters.

The difference between PreRequestParams and RequestParams is the absense of a base url to send requests to. When requests are load balanced between multiple Elasticsearch nodes the url to send a request to might not be known upfront.

Methods

impl PreRequestParams

pub fn new() -> Self

Create a new container for request parameters.

This method takes a fully-qualified url for the Elasticsearch node. It will also set the Content-Type header to application/json.

pub fn url_param<T: ToString>(self, key: &'static str, value: T) -> Self

Set a url param value.

These parameters are added as query parameters to request urls.

pub fn header<H>(self, header: H) -> Self where
H: Header + Clone,

Set a request header.