Struct elastic::client::RequestBuilder

pub struct RequestBuilder<'a, I, B> { /* fields omitted */ }

A builder for a request.

This structure wraps up a concrete REST API request type and lets you adjust parameters before sending it.

Methods

impl<'a, I, B> RequestBuilder<'a, I, B> where
I: Into<HttpRequest<'static, B>>,
B: IntoBody,

fn params<F>(self, builder: F) -> Self where
F: Fn(RequestParams) -> RequestParams,

Override the parameters for this request.

This method will clone the RequestParams on the Client and pass them to the closure.

Examples

Add a url param to force an index refresh:

client.request(get_req())
      .params(|params| params.url_param("refresh", true))
      .send()
      .unwrap();

fn send(self) -> Result<ResponseBuilder>

Send this request and return the response.

This method consumes the RequestBuilder and returns a ResponseBuilder that can be used to parse the response.