Module elastic::http

Raw HTTP modules.

These types are re-exported from reqwest and used in parts of elastics public API. They may eventually be wrapped and made implementation details.

Modules

header
Headers container, and common header fields.

Macros

header

Structs

AsyncBody
A type that can be converted into a request body.

AsyncBodyReader
A read adapter for an AsyncBody.

AsyncChunk
A chunk of bytes for a Body.

AsyncHttpResponse
A raw HTTP response that can be buffered using Read.

HttpRequest
A request just before being sent.

Method
The Request Method (VERB)

StatusCode
An HTTP status code (status-code in RFC 7230 et al.).

SyncBody
A type that can be converted into a request body.

SyncBodyReader
A read adapter for a SyncBody.

SyncHttpResponse
A raw HTTP response that can be buffered using Read.

Url
A parsed URL record.

Type Definitions

AsyncHttpRequest
A http request with an asynchronous body;

SyncHttpRequest
A http request with a synchronous body.