Struct elastic::types::date::DateValue

pub struct DateValue(_);

A date value produced and consumed by date formats.

DateValue is a very thin wrapper over DateTime<Utc> that doesn't carry any formatting semantics. Like FormattableDateValue, this type is used for binding generics in methods that accept date values but it ignores any format on the input type. You probably won't need to use it directly except to clobber the format on a Date<M> or DateTime<Utc> value.

Methods

impl DateValue

pub fn now() -> DateValue

Equivalent to DateTime<Utc>::now()

pub fn build(
year: i32,
month: u32,
day: u32,
hour: u32,
minute: u32,
second: u32,
milli: u32
) -> DateValue

Construct a DateValue from individual parts.