Module elastic::types::number

Implementation of the Elasticsearch number types.

Numeric types come in a number of flavours that correspond to primitives in Rust:

Rust 	 Elasticsearch

i64 	 long

i32 	 integer

i16 	 short

i8 	 byte

f64 	 double

f32 	 float

For mapping a number with the default mapping, you can use the Rust primitive. If you need to use a custom mapping, then there is an Elastic* type for each number.

Examples

For defining your own number mapping, see mapping details.

Map with a default number (integer in this case):

struct MyType {
    pub field: i32
}

Map with a custom number (integer in this case):

struct MyType {
    pub field: Integer<MyIntegerMapping>
}

Map a custom type as a number field (integer in this case):

#[derive(Serialize)]
struct MyIntegerField(i32);

impl IntegerFieldType<DefaultIntegerMapping> for MyIntegerField {}

Links

Elasticsearch Doc

Modules

mapping
Mapping for the Elasticsearch number types.

prelude
Includes all types for the number type.

Structs

Byte
Number type with a given mapping.

Double
Number type with a given mapping.

Float
Number type with a given mapping.

Integer
Number type with a given mapping.

Long
Number type with a given mapping.

Short
Number type with a given mapping.