Module elastic::types::number::prelude

Includes all types for the number type.

This is a convenience module to make it easy to build mappings for multiple types without too many use statements.

Structs

Byte
Number type with a given mapping.

DefaultByteMapping
Default mapping for a byte type.

DefaultDoubleMapping
Default mapping for a double type.

DefaultFloatMapping
Default mapping for a float type.

DefaultIntegerMapping
Default mapping for an integer type.

DefaultLongMapping
Default mapping for a long type.

DefaultShortMapping
Default mapping for a short type.

Double
Number type with a given mapping.

Float
Number type with a given mapping.

Integer
Number type with a given mapping.

Long
Number type with a given mapping.

Short
Number type with a given mapping.

Traits

ByteFieldType
A field that will be mapped as a number.

ByteMapping
Base number mapping.

DoubleFieldType
A field that will be mapped as a number.

DoubleMapping
Base number mapping.

FloatFieldType
A field that will be mapped as a number.

FloatMapping
Base number mapping.

IntegerFieldType
A field that will be mapped as a number.

IntegerMapping
Base number mapping.

LongFieldType
A field that will be mapped as a number.

LongMapping
Base number mapping.

ShortFieldType
A field that will be mapped as a number.

ShortMapping
Base number mapping.