Trait elastic::types::number::prelude::DoubleMapping

pub trait DoubleMapping: Default + Default {
    fn coerce() -> Option<bool> { ... }

    fn boost() -> Option<f32> { ... }

    fn doc_values() -> Option<bool> { ... }

    fn ignore_malformed() -> Option<bool> { ... }

    fn include_in_all() -> Option<bool> { ... }

    fn index() -> Option<bool> { ... }

    fn null_value() -> Option<f64> { ... }

    fn store() -> Option<bool> { ... }
}

Base number mapping.

Provided Methods

fn coerce() -> Option<bool>

Try to convert strings to numbers and truncate fractions for integers. Accepts true (default) and false.

fn boost() -> Option<f32>

Field-level index time boosting. Accepts a floating point number, defaults to 1.0.

fn doc_values() -> Option<bool>

Should the field be stored on disk in a column-stride fashion, so that it can later be used for sorting, aggregations, or scripting? Accepts true (default) or false.

fn ignore_malformed() -> Option<bool>

If true, malformed numbers are ignored. If false (default), malformed numbers throw an exception and reject the whole document.

fn include_in_all() -> Option<bool>

Whether or not the field value should be included in the _all field? Accepts true or false. Defaults to false if index is set to no, or if a parent object field sets include_in_all to false. Otherwise defaults to true.

fn index() -> Option<bool>

Should the field be searchable? Accepts not_analyzed (default) and no.

fn null_value() -> Option<f64>

Accepts a numeric value of the same type as the field which is substituted for any explicit null values. Defaults to null, which means the field is treated as missing.

fn store() -> Option<bool>

Whether the field value should be stored and retrievable separately from the _source field. Accepts true or false (default).