Module elastic::types::prelude

Includes all data types.

This is a convenience module to make it easy to build mappings for multiple types without too many use statements.

Structs

BasicDateTime
Format for basic_date_time.

BasicDateTimeNoMillis
Format for basic_date_time_no_millis.

Boolean
An Elasticsearch boolean with a mapping.

Byte
Number type with a given mapping.

ChronoFormat
Format for default chrono::DateTime.

Date
An Elasticsearch date type with a required time component.

DefaultBooleanMapping
Default mapping for bool.

DefaultByteMapping
Default mapping for a byte type.

DefaultDateMapping
Default mapping for date.

DefaultDoubleMapping
Default mapping for a double type.

DefaultFloatMapping
Default mapping for a float type.

DefaultGeoPointMapping
Default mapping for geo_point.

DefaultGeoShapeMapping
Default mapping for geo_shape.

DefaultIntegerMapping
Default mapping for an integer type.

DefaultIpMapping
Default mapping for geo_shape.

DefaultKeywordMapping
Default mapping for bool.

DefaultLongMapping
Default mapping for a long type.

DefaultMapping
A mapping implementation for a non-core type, or anywhere it's ok for Elasticsearch to infer the mapping at index-time.

DefaultShortMapping
Default mapping for a short type.

DefaultStringMapping
Default mapping for String.

DefaultTextMapping
Default mapping for bool.

Distance
A distance value paired with a unit of measure.

Document
A wrapper type for serialising user types.

Double
Number type with a given mapping.

ElasticCompletionFieldMapping
A multi-field string mapping for a completion suggester.

ElasticTokenCountFieldMapping
A multi-field string mapping for a token count.

EpochMillis
Format for epoch_millis.

Field
A wrapper type used to work around conflicting implementations of Serialize for the various mapping traits.

FieldDataFrequencyFilter
Fielddata for term frequency as a percentage range.

Float
Number type with a given mapping.

GeoPoint
An Elasticsearch geo_point type with a format.

GeoPointArray
Geo-point expressed as an array with the format: [lon, lat]

GeoPointHash
Geo-point expressed as a geohash.

GeoPointObject
Geo-point expressed as an object, with lat and lon keys.

GeoPointString
Geo-point expressed as a string with the format: "lat,lon".

GeoShape
Geo shape type with a given mapping.

Integer
Number type with a given mapping.

Ip
An Elasticsearch ip with a mapping.

Keyword
An Elasticsearch keyword with a mapping.

KeywordFieldMapping
A multi-field string mapping.

Long
Number type with a given mapping.

ParseError
Represents an error encountered during parsing.

Short
Number type with a given mapping.

Text
An Elasticsearch text field with a mapping.

TextFieldMapping
A multi-field string mapping.

WrappedMapping
Mapping for a wrapped value, like an array or optional type.

Enums

DistanceUnit
A unit of measure for distance.

Dynamic
The dynamic setting may be set at the mapping type level, and on each inner object. Inner objects inherit the setting from their parent object or from the mapping type.

IndexAnalysis
Should the field be searchable? Accepts not_analyzed (default) and no.

IndexOptions
The index_options parameter controls what information is added to the inverted index, for search and highlighting purposes.

Orientation
This parameter defines one of two coordinate system rules (Right-hand or Left-hand) each of which can be specified in a few different ways.

Strategy
The strategy defines the approach for how to represent shapes at indexing and search time.

StringField
A string sub-field type.

TermVector
Term vectors contain information about the terms produced by the analysis process.

Tree
Name of the PrefixTree implementation to be used.

Constants

BOOLEAN_DATATYPE
Elasticsearch datatype name.

BYTE_DATATYPE
Elasticsearch datatype name.

COMPLETION_DATATYPE
Elasticsearch datatype name.

DATE_DATATYPE
Elasticsearch datatype name.

DOUBLE_DATATYPE
Elasticsearch datatype name.

DYNAMIC_DATATYPE
Elasticsearch datatype name.

FLOAT_DATATYPE
Elasticsearch datatype name.

GEOPOINT_DATATYPE
Elasticsearch datatype name.

GEOSHAPE_DATATYPE
Elasticsearch datatype name.

INTEGER_DATATYPE
Elasticsearch datatype name.

IP_DATATYPE
Elasticsearch datatype name.

KEYWORD_DATATYPE
Elasticsearch datatype name.

LONG_DATATYPE
Elasticsearch datatype name.

NESTED_DATATYPE
Elasticsearch datatype name.

OBJECT_DATATYPE
Elasticsearch datatype name.

SHORT_DATATYPE
Elasticsearch datatype name.

TEXT_DATATYPE
Elasticsearch datatype name.

TOKENCOUNT_DATATYPE
Elasticsearch datatype name.

Traits

BooleanMapping
The base requirements for mapping a boolean type.

ByteMapping
Base number mapping.

DateFormat
A format used for parsing and formatting dates.

DateMapping
The base requirements for mapping a date type.

Datelike
The common set of methods for date component.

DocumentMapping
The base requirements for mapping an object type.

DocumentType
The additional fields available to an indexable Elasticsearch type.

DoubleMapping
Base number mapping.

FieldMapping
The base requirements for mapping an Elasticsearch data type.

FieldType
The base representation of an Elasticsearch data type.

FloatMapping
Base number mapping.

GeoPointFormat
A format used for parsing and formatting geo points.

GeoPointMapping
The base requirements for mapping a geo_point type.

GeoShapeMapping
The base requirements for mapping a geo_shape type.

IntegerMapping
Base number mapping.

IpMapping
The base requirements for mapping a ip type.

KeywordMapping
The base requirements for mapping a string type.

LongMapping
Base number mapping.

PropertiesMapping
Serialisation for the mapping of object properties.

SerializeField
A serialisable variant of the field mapping.

ShortMapping
Base number mapping.

TextMapping
The base requirements for mapping a string type.

Timelike
The common set of methods for time component.

Functions

doc_ser
Serialise a document mapping using the given serialiser.

field_ser
Serialise a field mapping using the given serialiser.

format_with_tokens
Format a date string using an owned slice of items.

mapping
Get the mapping for a field.

parse_from_tokens
Parse a date string using an owned slice of items.

Type Definitions

DefaultDateFormat
The default date format (BasicDateTime).

DefaultGeoPointFormat
The default geo_point format (GeoPointArray).