Struct elastic::types::prelude::Field

pub struct Field<M, F> where
    F: Default,
    M: FieldMapping<F>,  { /* fields omitted */ }

A wrapper type used to work around conflicting implementations of Serialize for the various mapping traits.

Serialising Field will produce the mapping for the given type, suitable as the mapping of a field for a document.