Struct elastic::types::prelude::KeywordFieldMapping

pub struct KeywordFieldMapping {
    pub analyzer: Option<&'static str>,
    pub doc_values: Option<bool>,
    pub eager_global_ordinals: Option<bool>,
    pub include_in_all: Option<bool>,
    pub ignore_above: Option<u32>,
    pub index: Option<bool>,
    pub index_options: Option<IndexOptions>,
    pub norms: Option<bool>,
    pub store: Option<bool>,
    pub search_analyzer: Option<&'static str>,
    pub similarity: Option<&'static str>,
}

A multi-field string mapping.

Fields
analyzer: Option<&'static str>

The analyzer which should be used for analyzed string fields, both at index-time and at search-time (unless overridden by the search_analyzer). Defaults to the default index analyzer, or the standard analyzer.

doc_values: Option<bool>

Should the field be stored on disk in a column-stride fashion, so that it can later be used for sorting, aggregations, or scripting? Accepts true (default) or false.

eager_global_ordinals: Option<bool>

Should global ordinals be loaded eagerly on refresh? Accepts true or false (default). Enabling this is a good idea on fields that are frequently used for (significant) terms aggregations.

include_in_all: Option<bool>

Whether or not the field value should be included in the _all field? Accepts true or false. Defaults to false if index is set to no, or if a parent object field sets include_in_all to false. Otherwise defaults to true.

ignore_above: Option<u32>

The maximum number of characters to index. Any characters over this length will be ignored.

index: Option<bool>

Should the field be searchable? Accepts true (default) or false.

index_options: Option<IndexOptions>

What information should be stored in the index, for search and highlighting purposes. Defaults to Positions.

norms: Option<bool>

Whether field-length should be taken into account when scoring queries. Accepts true (default) or false.

store: Option<bool>

Whether the field value should be stored and retrievable separately from the _source field. Accepts true or false (default).

search_analyzer: Option<&'static str>

The analyzer that should be used at search time on analyzed fields. Defaults to the analyzer setting.

similarity: Option<&'static str>

Which scoring algorithm or similarity should be used. Defaults to "classic", which uses TF/IDF.