Struct elastic::types::string::mapping::ElasticCompletionFieldMapping

pub struct ElasticCompletionFieldMapping {
    pub analyzer: Option<&'static str>,
    pub search_analyzer: Option<&'static str>,
    pub payloads: Option<bool>,
    pub preserve_separators: Option<bool>,
    pub preserve_position_increments: Option<bool>,
    pub max_input_length: Option<u32>,
}

A multi-field string mapping for a completion suggester.

Fields
analyzer: Option<&'static str>

The analyzer which should be used for analyzed string fields, both at index-time and at search-time (unless overridden by the search_analyzer). Defaults to the default index analyzer, or the standard analyzer.

search_analyzer: Option<&'static str>

The search analyzer to use, defaults to value of analyzer.

payloads: Option<bool>

Enables the storing of payloads, defaults to false.

preserve_separators: Option<bool>

Preserves the separators, defaults to true. If disabled, you could find a field starting with Foo Fighters, if you suggest for foof.

preserve_position_increments: Option<bool>

Enables position increments, defaults to true. If disabled and using stopwords analyzer, you could get a field starting with The Beatles, if you suggest for b.

Note: You could also achieve this by indexing two inputs, Beatles and The Beatles, no need to change a simple analyzer, if you are able to enrich your data.

max_input_length: Option<u32>

Limits the length of a single input, defaults to 50 UTF-16 code points. This limit is only used at index time to reduce the total number of characters per input string in order to prevent massive inputs from bloating the underlying datastructure. The most usecases won’t be influenced by the default value since prefix completions hardly grow beyond prefixes longer than a handful of characters. (Old name "max_input_len" is deprecated)