Struct elastic::types::string::text::mapping::TextFieldMapping

pub struct TextFieldMapping {
    pub analyzer: Option<&'static str>,
    pub eager_global_ordinals: Option<bool>,
    pub fielddata: Option<bool>,
    pub fielddata_frequency_filter: Option<FieldDataFrequencyFilter>,
    pub include_in_all: Option<bool>,
    pub ignore_above: Option<u32>,
    pub index: Option<bool>,
    pub index_options: Option<IndexOptions>,
    pub norms: Option<bool>,
    pub position_increment_gap: Option<u32>,
    pub store: Option<bool>,
    pub search_analyzer: Option<&'static str>,
    pub search_quote_analyzer: Option<&'static str>,
    pub similarity: Option<&'static str>,
    pub term_vector: Option<TermVector>,
}

A multi-field string mapping.

Fields
analyzer: Option<&'static str>

The analyzer which should be used for analyzed string fields, both at index-time and at search-time (unless overridden by the search_analyzer). Defaults to the default index analyzer, or the standard analyzer.

eager_global_ordinals: Option<bool>

Should global ordinals be loaded eagerly on refresh? Accepts true or false (default). Enabling this is a good idea on fields that are frequently used for (significant) terms aggregations.

fielddata: Option<bool>

Can the field use in-memory fielddata for sorting, aggregations, or scripting? Accepts true or false (default).

fielddata_frequency_filter: Option<FieldDataFrequencyFilter>

Expert settings which allow to decide which values to load in memory when fielddata is enabled. By default all values are loaded.

include_in_all: Option<bool>

Whether or not the field value should be included in the _all field? Accepts true or false. Defaults to false if index is set to no, or if a parent object field sets include_in_all to false. Otherwise defaults to true.

ignore_above: Option<u32>

The maximum number of characters to index. Any characters over this length will be ignored.

index: Option<bool>

Should the field be searchable? Accepts true (default) or false.

index_options: Option<IndexOptions>

What information should be stored in the index, for search and highlighting purposes. Defaults to Positions.

norms: Option<bool>

Whether field-length should be taken into account when scoring queries. Accepts true (default) or false.

position_increment_gap: Option<u32>

The number of fake term position which should be inserted between each element of an array of strings. Defaults to the position_increment_gap configured on the analyzer which defaults to 100. 100 was chosen because it prevents phrase queries with reasonably large slops (less than 100) from matching terms across field values.

store: Option<bool>

Whether the field value should be stored and retrievable separately from the _source field. Accepts true or false (default).

search_analyzer: Option<&'static str>

The analyzer that should be used at search time on analyzed fields. Defaults to the analyzer setting.

search_quote_analyzer: Option<&'static str>

The analyzer that should be used at search time when a phrase is encountered. Defaults to the search_analyzer setting.

similarity: Option<&'static str>

Which scoring algorithm or similarity should be used. Defaults to "classic", which uses TF/IDF.

term_vector: Option<TermVector>

Whether term vectors should be stored for an analyzed field. Defaults to No.