Module elastic::client::requests

Request types for the Elasticsearch REST API.

This module contains implementation details that are useful if you want to customise the request process, but aren't generally important for sending requests.

Re-exports

pub use self::raw::RawRequestBuilder;

pub use self::search::SearchRequestBuilder;

pub use self::document_get::GetRequestBuilder;

pub use self::document_index::IndexRequestBuilder;

pub use self::document_update::UpdateRequestBuilder;

pub use self::document_delete::DeleteRequestBuilder;

pub use self::document_put_mapping::PutMappingRequestBuilder;

pub use self::index_create::IndexCreateRequestBuilder;

pub use self::index_open::IndexOpenRequestBuilder;

pub use self::index_close::IndexCloseRequestBuilder;

pub use self::index_delete::IndexDeleteRequestBuilder;

pub use self::index_exists::IndexExistsRequestBuilder;

pub use self::ping::PingRequestBuilder;

pub use self::bulk::BulkRequestBuilder;

Modules

bulk
Builders for bulk requests.

common
Types that are common between requests.

document_delete
Builders for delete document requests.

document_get
Builders for get document requests.

document_index
Builders for index requests.

document_put_mapping
Builders for put mapping requests.

document_update
Builders for update document requests.

endpoints
REST API endpoints.

index_close
Builders for close index requests.

index_create
Builders for create index requests.

index_delete
Builders for delete index requests.

index_exists
Builders for index exists requests.

index_open
Builders for open index requests.

params
Common url params like Id and Index.

ping
Builders for ping requests.

prelude
A glob import for convenience.

raw
Builders for raw requests.

search
Builders for search requests.

Structs

Alias

BulkRequest
Post: /_bulk

CatAliasesRequest
Get: /_cat/aliases

CatAllocationRequest
Get: /_cat/allocation

CatCountRequest
Get: /_cat/count

CatFielddataRequest
Get: /_cat/fielddata

CatHealthRequest
Get: /_cat/health

CatHelpRequest
Get: /_cat

CatIndicesRequest
Get: /_cat/indices

CatMasterRequest
Get: /_cat/master

CatNodeattrsRequest
Get: /_cat/nodeattrs

CatNodesRequest
Get: /_cat/nodes

CatPendingTasksRequest
Get: /_cat/pending_tasks

CatPluginsRequest
Get: /_cat/plugins

CatRecoveryRequest
Get: /_cat/recovery

CatRepositoriesRequest
Get: /_cat/repositories

CatSegmentsRequest
Get: /_cat/segments

CatShardsRequest
Get: /_cat/shards

CatSnapshotsRequest
Get: /_cat/snapshots

CatTasksRequest
Get: /_cat/tasks

CatTemplatesRequest
Get: /_cat/templates

CatThreadPoolRequest
Get: /_cat/thread_pool

ClearScrollRequest
Delete: /_search/scroll/{scroll_id}

ClusterAllocationExplainRequest
Post: /_cluster/allocation/explain

ClusterGetSettingsRequest
Get: /_cluster/settings

ClusterHealthRequest
Get: /_cluster/health

ClusterPendingTasksRequest
Get: /_cluster/pending_tasks

ClusterPutSettingsRequest
Put: /_cluster/settings

ClusterRerouteRequest
Post: /_cluster/reroute

ClusterStateRequest
Get: /_cluster/state

ClusterStatsRequest
Get: /_cluster/stats

CountPercolateRequest
Post: /{index}/{type}/_percolate/count

CountRequest
Post: /_count

CreateRequest
Post: /{index}/{type}/{id}/_create

DeleteByQueryRequest
Post: /{index}/_delete_by_query

DeleteRequest
Delete: /{index}/{type}/{id}

DeleteScriptRequest
Delete: /_scripts/{lang}/{id}

DeleteTemplateRequest
Delete: /_search/template/{id}

Endpoint
A general request type that all endpoints can be converted into.

ExistsRequest
Head: /{index}/{type}/{id}

ExplainRequest
Post: /{index}/{type}/{id}/_explain

Feature

FieldStatsRequest
Post: /_field_stats

Fields

GetRequest
Get: /{index}/{type}/{id}

GetScriptRequest
Get: /_scripts/{lang}/{id}

GetSourceRequest
Get: /{index}/{type}/{id}/_source

GetTemplateRequest
Get: /_search/template/{id}

Id

Index

IndexMetric

IndexRequest
Post: /{index}/{type}

IndicesAnalyzeRequest
Post: /_analyze

IndicesClearCacheRequest
Post: /_cache/clear

IndicesCloseRequest
Post: /{index}/_close

IndicesCreateRequest
Put: /{index}

IndicesDeleteAliasRequest
Delete: /{index}/_alias/{name}

IndicesDeleteRequest
Delete: /{index}

IndicesDeleteTemplateRequest
Delete: /_template/{name}

IndicesExistsAliasRequest
Head: /_alias/{name}

IndicesExistsRequest
Head: /{index}

IndicesExistsTemplateRequest
Head: /_template/{name}

IndicesExistsTypeRequest
Head: /{index}/_mapping/{type}

IndicesFlushRequest
Post: /_flush

IndicesFlushSyncedRequest
Post: /_flush/synced

IndicesForcemergeRequest
Post: /_forcemerge

IndicesGetAliasRequest
Get: /_alias/

IndicesGetFieldMappingRequest
Get: /_mapping/field/{fields}

IndicesGetMappingRequest
Get: /_mapping

IndicesGetRequest
Get: /{index}

IndicesGetSettingsRequest
Get: /_settings

IndicesGetTemplateRequest
Get: /_template/{name}

IndicesGetUpgradeRequest
Get: /_upgrade

IndicesOpenRequest
Post: /{index}/_open

IndicesPutAliasRequest
Post: /{index}/_alias/{name}

IndicesPutMappingRequest
Post: /{index}/{type}/_mapping

IndicesPutSettingsRequest
Put: /_settings

IndicesPutTemplateRequest
Post: /_template/{name}

IndicesRecoveryRequest
Get: /_recovery

IndicesRefreshRequest
Post: /_refresh

IndicesRolloverRequest
Post: /{alias}/_rollover

IndicesSegmentsRequest
Get: /_segments

IndicesShardStoresRequest
Get: /_shard_stores

IndicesShrinkRequest
Post: /{index}/_shrink/{target}

IndicesStatsRequest
Get: /_stats

IndicesUpdateAliasesRequest
Post: /_aliases

IndicesUpgradeRequest
Post: /_upgrade

IndicesValidateQueryRequest
Post: /_validate/query

InfoRequest
Get: /

IngestDeletePipelineRequest
Delete: /_ingest/pipeline/{id}

IngestGetPipelineRequest
Get: /_ingest/pipeline/{id}

IngestPutPipelineRequest
Put: /_ingest/pipeline/{id}

IngestSimulateRequest
Post: /_ingest/pipeline/_simulate

Lang

Metric

MgetRequest
Post: /_mget

MpercolateRequest
Post: /_mpercolate

MsearchRequest
Post: /_msearch

MsearchTemplateRequest
Post: /_msearch/template

MtermvectorsRequest
Post: /_mtermvectors

Name

NewIndex

NodeId

NodesHotThreadsRequest
Get: /_nodes/hot_threads

NodesInfoRequest
Get: /_nodes

NodesStatsRequest
Get: /_nodes/stats

PercolateRequest
Post: /{index}/{type}/_percolate

PingHeadRequest
Head: /

PingRequest
Get: /

PutScriptRequest
Post: /_scripts/{lang}/{id}

PutTemplateRequest
Post: /_search/template/{id}

ReindexRequest
Post: /_reindex

ReindexRethrottleRequest
Post: /_reindex/{task_id}/_rethrottle

RenderSearchTemplateRequest
Post: /_render/template

Repository

RequestBuilder
A builder for a request.

ScrollId

ScrollRequest
Post: /_search/scroll

SearchRequest
Post: /_search

SearchShardsRequest
Post: /{index}/_search_shards

SearchTemplateRequest
Post: /_search/template

SimpleSearchRequest
Get: /_search

Snapshot

SnapshotCreateRepositoryRequest
Post: /_snapshot/{repository}

SnapshotCreateRequest
Post: /_snapshot/{repository}/{snapshot}

SnapshotDeleteRepositoryRequest
Delete: /_snapshot/{repository}

SnapshotDeleteRequest
Delete: /_snapshot/{repository}/{snapshot}

SnapshotGetRepositoryRequest
Get: /_snapshot

SnapshotGetRequest
Get: /_snapshot/{repository}/{snapshot}

SnapshotRestoreRequest
Post: /_snapshot/{repository}/{snapshot}/_restore

SnapshotStatusRequest
Get: /_snapshot/_status

SnapshotVerifyRepositoryRequest
Post: /_snapshot/{repository}/_verify

SuggestRequest
Post: /_suggest

Target

TaskId

TasksCancelRequest
Post: /_tasks

TasksGetRequest
Get: /_tasks/{task_id}

TasksListRequest
Get: /_tasks

TermvectorsRequest
Post: /{index}/{type}/_termvectors

ThreadPoolPatterns

Type

UpdateByQueryRequest
Post: /{index}/_update_by_query

UpdateRequest
Post: /{index}/{type}/{id}/_update

UrlPath
A wrapper around an owned or borrowed url path.

Functions

alias

empty_body
A convenience method for a default, empty body. This method doesn't allocate.

feature

fields

id

index

index_metric

lang

metric

name

new_index

node_id

repository

scroll_id

snapshot

target

task_id

thread_pool_patterns

ty

Type Definitions

DefaultBody
A default body type.