Module elastic::client::responses

Response types for the Elasticsearch REST API.

All high-level builders return one of these response types for you. This module also contains helpers that can be used to parse responses from a raw HttpResponse.

Modules

bulk
Response types for a bulk request.

parse
Utility types for response parsing.

prelude
A glob import for convenience.

search
Response types for a search request.

Structs

AsyncResponseBuilder
A builder for a response.

BulkErrorsResponse
Response for a bulk request.

BulkResponse
Response for a bulk request.

CommandResponse
A standard command acknowledgement response.

DeleteResponse
Response for a delete document request.

GetResponse
Response for a get document request.

IndexResponse
Response for an index document request.

IndicesExistsResponse
Response for an index exists request.

IntoResponse
A future returned by calling into_response.

PingResponse
Response for a cluster ping request.

SearchResponse
Response for a search request.

Shards
Returned hits metadata.

SyncResponseBuilder
A builder for a response.

UpdateResponse
Response for a update document request.