Module elastic::client::sender

Low-level http client details.

This module contains low-level implementation details for the high-level Client. Most of the types here aren't currently usable directly, but once the low-level API matures they should be extensible. Some notable types include:

Sender: a generic trait that can send a http request and return a response

NextParams: a generic trait that can fetch a set of parameters to associate with a request

SyncSender: a synchronous http client

AsyncSender: an asynchronous http client.

Modules

sniffed_nodes
Load balanced nodes sniffed regularly from an Elasticsearch cluster.

static_nodes
Multiple static nodes that can be load balanced by some strategy.

Structs

AsyncClientBuilder
A builder for an asynchronous client.

AsyncSender
An asynchronous request sender.

NodeAddress
A single node address.

NodeAddresses
A common container for a source of node addresses.

Params
A set of parameters returned by calling next on a sync set of NodeAddresses.

PendingParams
A future returned by calling next on an async set of NodeAddresses.

PendingResponse
A future returned by calling send on an AsyncSender.

PreRequestParams
An incomplete set of request parameters.

RequestParams
Parameters for a single REST API request.

SendableRequest
A sendable request.

SyncClientBuilder
A builder for a syncronous client.

SyncSender
A synchronous request sender.

Constants

DEFAULT_NODE_ADDRESS

Traits

IntoAsyncHttpClient
A type that can be used to construct an async http client.

NextParams
Represents a type that can fetch request parameters.

Sender
Represents a type that can send a request.

Type Definitions

AsyncClient
An asynchronous Elasticsearch client.

SyncClient
A synchronous Elasticsearch client.