Struct elastic::client::sender::SyncClientBuilder

pub struct SyncClientBuilder { /* fields omitted */ }

A builder for a syncronous client.

Methods

impl SyncClientBuilder

pub fn new() -> Self

Create a new client builder.

By default, a client constructed by this builder will:

Send requests to localhost:9200

Not use any authentication

Not use TLS

pub fn from_params(params: PreRequestParams) -> Self

Create a new client builder with the given default request parameters.

pub fn static_node<S>(self, node: S) -> Self where
S: Into<NodeAddress>,

Specify a static node nodes to send requests to.

pub fn static_nodes<I, S>(self, nodes: I) -> Self where
I: IntoIterator<Item = S>,
S: Into<NodeAddress>,

Specify a set of static node nodes to load balance requests on.

pub fn sniff_nodes<I>(self, builder: I) -> Self where
I: Into<SniffedNodesBuilder>,

Specify a node address to sniff other nodes in the cluster from.

Examples

Use a given base url for sniffing the cluster's node addresses from:

let builder = SyncClientBuilder::new()
    .sniff_nodes("http://localhost:9200");

pub fn sniff_nodes_fluent<I, F>(self, address: I, builder: F) -> Self where
I: Into<NodeAddress>,
F: Fn(SniffedNodesBuilder) -> SniffedNodesBuilder + 'static,

Specify a node address to sniff other nodes in the cluster from.

Examples

Use a given base url for sniffing the cluster's node addresses from and specify a minimum duration to wait before refreshing:

let builder = SyncClientBuilder::new()
    .sniff_nodes_fluent("http://localhost:9200", |n| n
        .wait(Duration::from_secs(90)));

pub fn params_fluent<F>(self, builder: F) -> Self where
F: Fn(PreRequestParams) -> PreRequestParams + 'static,

Specify default request parameters.

Examples

Require all responses use pretty-printing:

let builder = SyncClientBuilder::new()
    .params_fluent(|p| p
        .url_param("pretty", true));

Add an authorization header:

use elastic::http::header::Authorization;

let builder = SyncClientBuilder::new()
    .params_fluent(|p| p
        .header(Authorization("let me in".to_owned())));

pub fn params<I>(self, params: I) -> Self where
I: Into<PreRequestParams>,

Specify default request parameters.

Examples

Require all responses use pretty-printing:

let builder = SyncClientBuilder::new()
    .params(PreRequestParams::default()
        .url_param("pretty", true));

Add an authorization header:

use elastic::http::header::Authorization;

let builder = SyncClientBuilder::new()
    .params(PreRequestParams::default()
        .header(Authorization("let me in".to_owned())));

pub fn http_client(self, client: SyncHttpClient) -> Self

Use the given reqwest::Client for sending requests.

pub fn pre_send_raw<F>(self, pre_send: F) -> Self where
F: Fn(&mut SyncHttpRequest) -> Result<(), Box<dyn StdError + Send + Sync>> + Send + Sync + 'static,

Specify a function to tweak a raw request before sending.

This function will be applied to all outgoing requests and gives you the chance to perform operations the require the complete raw request, such as request singing. Prefer the params method on the client or individual requests where possible.

pub fn build(self) -> Result<SyncClient, Error>

Construct a SyncClient from this builder.