Module elastic::http

Raw HTTP modules.

These types are re-exported from reqwest and used in parts of elastics public API. They may eventually be wrapped and made implementation details.

Modules

header
Headers container, and common header fields.

Structs

AsyncChunk
A chunk of bytes for a Body.

SyncBody
The body of a Request.