Struct elastic::types::date::prelude::Date

pub struct Date<TMapping> where
    TMapping: DateMapping,  { /* fields omitted */ }

An Elasticsearch date type with a required time component.

The format is provided as a generic parameter. This struct wraps up a chrono::DateTime<Utc> struct, meaning storing time in Utc is required.

Examples

Defining a date using the default format:

let date: Date<DefaultDateMapping> = Date::now();

Defining a date using a named format:

let date: Date<DefaultDateMapping<BasicDateTime>> = Date::now();

Accessing the values of a date:

let date = Date::<DefaultDateMapping>::now();

//eg: 2010/04/30 13:56:59.372
println!("{}/{}/{} {}:{}:{}.{}",
    date.year(),
    date.month(),
    date.day(),
    date.hour(),
    date.minute(),
    date.second(),
    date.nanosecond() / 1000000
);

Links

Elasticsearch Doc

Methods

impl<TMapping> Date<TMapping> where
TMapping: DateMapping,

pub fn new<I>(date: I) -> Date<TMapping> where
I: Into<FormattableDateValue<<TMapping as DateMapping>::Format>>,

Creates a new Date from the given chrono::DateTime<Utc>.

This function will consume the provided chrono date.

Examples

Create a Date from a DateValue:

//Create a DateValue struct
let date = DateValue::now();
    
//Give it to the Date struct
let date: Date<DefaultDateMapping> = Date::new(date);

If the Dates format is ChronoFormat, then it can also be created from chrono::DateTime:

use chrono::Utc;
    
//Create a chrono DateTime struct
let chronoDate = Utc::now();
    
//Give it to the Date struct
let date: Date<DefaultDateMapping<ChronoFormat>> = Date::new(chronoDate);

If the Dates format isn't ChronoFormat, then the chrono::DateTime will need to be converted into a DateValue first. This is to make sure you don't accidentally change the format of a date, which could lead to errors at runtime:

use chrono::Utc;
    
//Create a chrono DateTime struct
let chronoDate = Utc::now();
    
//Give it to the Date struct
let date: Date<DefaultDateMapping<EpochMillis>> = Date::new(DateValue::from(chronoDate));

pub fn build(
year: i32,
month: u32,
day: u32,
hour: u32,
minute: u32,
second: u32,
milli: u32
) -> Date<TMapping>

Creates an Date from the given Utc primitives:

let date: Date<DefaultDateMapping> = Date::build(2015, 5, 14, 16, 45, 8, 886);

pub fn now() -> Date<TMapping>

Gets the current system time.

Examples

let date: Date<DefaultDateMapping> = Date::now();

pub fn remap<TNewMapping>(date: Date<TMapping>) -> Date<TNewMapping> where
TNewMapping: DateMapping,

Change the format/mapping of this date.

Examples

//Get the current datetime formatted as basic_date_time
let date: Date<DefaultDateMapping<BasicDateTime>> = Date::now();
    
//Change the format to epoch_millis
let otherdate: Date<DefaultDateMapping<EpochMillis>> = Date::remap(date);